Corrects the behavior of `uavcan::tx::breakdown::crc_kind`.
`crc_kind` is tasked with identifying how the crc of a multi frame transfer
should be positioned.

The implementation before this commit was broken, wrongly identifying or failing
to identify some of the possible cases, panicking in the process.

For example, given an `MTU` and a payload of length `N(MTU-1) + 1`, where `N` is a positive
integer, `crc_kind` panicked instead of identifying the case as an
`CRCKind::Embedded` case.

This behavior depended by the incorrect specification of the conditions inside
`crc_kind` branching logic.

The behavior is now corrected.

Further to correcting the behavior, a series of dictionary changes were made to
better describe the practical meaning of the `crc_kind` function.
In particular, its only argument was renamed to `payload_reminder` from
`last_frame_data_len`.

`Breakdown` uses `chunks_exact` to cut the payload into chunks of the correct
length to be sent in a frame. The length of the remainder of chunks exact is
what is passed to `crc_kind` when `Breakdown` calls it.

This fixes the presence of a check, in `crc_kind`, for being passed an argument
with the `0` value.

`last_frame_data_len` identifies an argument that contains the amount of data in
the last frame of the transfer or the last produced frame of `Breakdown`.
When there is no data, that is, a value of `0` inhabits `last_frame_data_len`,
then no frame should exist at all, making `last_frame_data_len` paradoxically
incorrect.

Nonetheless, a value of `0` may be encountered, and must be handled, with the
current implementation.
Hence, the new naming which correctly describes the meaning of the argument of
`crc_kind` and justifies its ability to be `0` valued.