Realigns the `hello_can` examples with the current interface.
The `CanFrame` boilerplate was modified to correctly handle non-full frames
during the conversions between `socketcan` frames and internally used frames.

This required the realignment with the new interface, where the length is passed
around during the conversions.
Furthermore, the write implementation for the `Can` wrapper was modified to copy
only the correct amount of bytes.