Aligns the tail byte tests to the specification.
`the_successor_of_a_tail_byte_has_a_transfer_id_that_is_the_successor_of_the_transfer_id_of_the_original_tail_byte`
test was testing for a behavior that does not respect the uavcan specification.

This behavior is a refuse of a misunderstanding of the specification and an
incorrect implementation.

Indeed, transfer-ids are static during a multi-frame transfer.

The behavior was previously aligned with a corrected implementation but the test
was not changed, resulting in the incorrect behavior being tested and test
failing fallaciously.

The test is now aligned with the implementation and tests the correct behavior.