Removes the tests::common::data_generation module
Previously, tests were entirely done in the tests directory.
Later, they were divided into module-level unit-tests and integration-tests in
tests directory.

To allow both of the tests to use property tests without repetition, the code in
the data generation module, which previously contained the proptest strategy
definitions for the crate types, was moved locally into the module that defines
the type that the strategy generates.

Hence, the data_generation module is now removed as it is unused.