Removes the tests::common::padding module.
Initially, the library always positioned the tail byte at the end of a frame and
the crc on its own frame.
This means that, for example, if the last frame of the payload only occupied 2
bytes, 5 bytes of padding would have been added, followed by the tail byte.

This was later changed to respect the uavcan specification examples, where only
the required amount of bytes were sent and the tail byte always followed the
last byte of the transfer.

The `padding` module provided a simple function to calculate how many
padding bytes should be inserted, such that the expected reconstructed payload
could be built for equivalence testing when testing the transfer procedure.

Since it is not the case anymore that padding is required, the module is now
removed.

The
`receiving_the_frames_of_a_transmission_rebuilds_the_original_payload_along_with_the_padding_bytes`
was renamed to remove any reference to the padding bytes and its implementation
was changed to compare the original payload to the reconstructed one.