Converts SessionId TryFrom implementation to From.
`SessionId` is converted from a u32, representing the can id of the passed in
frame, during the `Buildup` process.

The `uavcan::session_id::session_id` module provided an implementation of
`TryFrom<u32>` for this reason.

At the level of `SessionId`, currently albeit this may change, any 32 bit value
is considered correct as the only discriminant is the value of the 26th bit,
which is always either zero or one.

For this reason, the implementation of `TryFrom` had no sensible way to return
an error and was converted to a `From` implementation.

The `TryFrom` implementation was further wrongly implemented, considering a
0-valued 26th bit both a message and a service, creating an unreachable branch,
and providing a defensive panic, which should have been unreachable, if this was
the case.

The new `From` implementation corrects the incorrect behavior and simplifies the
implementation.

The call to `SessionId::try_from` in `Buildup` was modified to call the new
`From` implementation.