Fixes SessionKind implementation of From<SessionId>.
`SessionKind::Response` contains a `Request` that should represent the original
request for which this response exists.

When a `Response` is converted to a can id in number form, as per the
specification, the source node id and destination node id of the original
request are inverted to represent the response.

This was not taken into account into the `From<SessionId>` implementation for
`SessionKind`, which restored the response's request with the two node ids
inverted.

The `From<SessionId>` implementation is now correctly swaps the two values to
reconstruct the original response.