Removes proptest from the non-dev dependencies.
The integration tests in the tests directory make use of some of the proptest
strategies defined in the main crate.

As proptest mainly needs std or alloc, the strategy modules and the tests are
intended to be compiled and used in a std environment only.

The strategy modules are shared between both the module-level tests and the
integration tests in the tests directory.

As they make use of std and the crate is intended to be used in a no_std
environment, they need to be compiled only when the crate is built for testing.

This is supported trough conditional compilation on `test` for module-level
tests.
For integration tests, rust compiles the crate with the deb profile and then links it
to the compiled test file.

With the dev profile, test is not active and cannot be conditionally compiled
upon.
Furthermore, dev-dependencies are not available, forcing the `proptest`
dependency to be available in normal-dependencies too.

Since the `proptest` dependency for the dev-profile with std support, debug
builds cannot be compiled anymore for a no_std environment.

Rust doesn't allow to specify dependencies specifically for integration tests
nor does it allow conditional compilation for integration tests only.

Most workarounds for this have undesirable tradeoffs.

For example, using a feature to condition for all test related code and for the
optional proptest dependency, make it so
`cargo test` will stop working, as there is no direct way to automatically add
the feature for test builds only, requiring a feature parameter to be passed to
`cargo test`.

Moving the tests to a different crate may work, but might require to not use the
test infrastructure directly or to lose access to unit testing private parts of
the crate.

Other solutions have shown to provide similar tradeoffs.

Since rust doesn't seem to support this use case directly, to avoid having to
build cruft around this limitation, the problem is removed at the root.

The tests directory was now removed, and is intended to be re-added only for tests
that do not depend on property testing.

The end-to-end tests that were in the tests directory, along with their
supporting code, were moved to module-level testing, where they have access to
their dependencies behind `cfg(test)`.

All the strategy modules are now conditioned on `test` only, as they are not
needed anymore with the `dev-profile`.

Finally, `proptest` is removed from `dependencies` and remain as a
`dev-dependency` only.