Hides dev-dependencies behind a target flag.
Cargo build, normally, downloads and compiles all dependencies independently of
if they are needed or not.

In the case of uavcan, this means that dev-dependencies, which depend on std and
are used only for tests and examples, are downloaded even when building only the
library targeting an embedded device, where std is not available, failing the
compilation.

To avoid this issue, all std-based dev-dependencies are now hidden behind a
target flag that comprises the general mainstream OSes.

Furthermore, the resolver for cargo is now set to default to the second version
to avoid downloading dependencies that are needed on specific platforms only
when building for a different target.
See
https://doc.rust-lang.org/cargo/reference/features.html#feature-resolver-version-2
for an explanation.