Adds an import for std::format in breakdown.rs' tests.
Proptest requires std::format to work correctly.

Since the crate is intended to be #![no_std], std is not externed by default and
its prelude is not being imported.

Since the prelude is not automatically imported, `std::format` is not available
for those tests that requires it even when `std` is externed.

The tests in `breakdown.rs` use proptest and require it, otherwise cargo test
will not be able to build correctly.

For this reason, `std::format` was added as an import in `breakdown.rs`' tests module.