Moves to a personal fork of modular-bitfield.
The modular bitfield crate produces an unneccesary braces warning when `From`
implementations are generated.

See: https://github.com/Robbepop/modular-bitfield/issues/66

A pull request ( https://github.com/Robbepop/modular-bitfield/pull/68 ) to
resolve this was provided.

Since it is not yet accepted mainstream, the `modular-bitfield` dependency was
temporarily moved to a custom fork to move forward with the removal of
compilation warnings.