Moves clippy::module_inception from warn to allow.
`clippy::module_inception` warns when a module has the same name as its
containing module.

For this library this happens in a few places.
For example, the top-level module directory `tail_byte` has a `tail_byte.rs`
file inside.

The style for this library dictates that, generally, the `mod.rs` files are used
principally to export the final interface, which is not yet done for each module
but will be in the future, while each small or central type gets its own file.

In this case, the top-level `tail_byte` module represents the `tail-byte`
concept from uavcan, while the internal `tail_byte` module from `tail_byte.rs`
represents the specific type that is used for the `tail_byte`.

This convention is in contrast with clippy defaults warns.

Since the library takes a standing that this is a cleaner solution,
`clippy::module_inception` was moved from `warn` to `allow`.