Switches from or to or_else when a function call is involved.
The clippy lint `or_fun_call` warns when the non-lazy version of the branching
chain-operators is used with a function call instead of the lazy version
`*or_else`.

Using the lazy version can slightly improve performance and avoid the case where
a side-effect is executed when it shouldn't.

For this reason, the codebase was modified to comply with the lint.