Removes unneeded closures in higher order functions.
Clippy's `redundant_closures` warns when a closure is created to wrap a function
call that has the same signature as the closure and can, hence, be called
directly by reference.

To comply with the lint and lighten the code, the codebase was modified to
remove the wrapping closures from those type of calls.