Switches definitely to the custom fork of modular_bitfield.
The custom fork was initially made as a temporary solution to the warning
problem in `modular_bitfield` while we waited for upstream PRs to be merged.

Since the `modular_bitfield` repository is moving slowly and other
warning-related problems have come up, the custom fork is now considered
permanent to allow for moving along with the issues.

For this reasons, the work on the custom fork was merged into its master branch
and the dependency in `uavcan` was modified to not use a custom branch anymore.

The latest changes in `modular_bitfield` further resolves the problem with
clippy's `identity_op` warning in autogenerated `From` implementations.