Renames *as_u32 to *from_base_parts.
The various `*as_u32` methods in `MessageSessionId` and `ServiceSessionId`,
where used to build the types from the other base types provided by the crate,
such as `TransferPriority`, and then to convert them to a `u32`, to be used in
places where the numeral representation was needed as a `can_id`.

In the specific case of `MessageSessionId::as_u32`, the naming clashed with
`clippy::wrong_self_convention`, as it started with `as` but took no self.

The naming was a workaround as `new` is unavailable for `modular_bitfield`
generated crate, conflicting with the autogenerated implementation.

To support both the `modular_bitfield` requirement and the clippy sensible lint,
the method was renamed to `from_base_parts`, to identify a construction of an
instance from the small, changing, parts that composes it.

Furthermore, to better align the method with a "constructor" feeling, the return
type was changed to `Self` and the conversion to `u32` was delegated to the user
of the method, which can use the autogenerated `From` implementation to obtain
the required representation.

All the places that previously used the method were updated with the new naming
and by surrounding the call in a `u32::from` expression.

To keep consistency, `ServiceSessionId::request_as_u32` and
`ServiceSessionId::response_as_u32` were modified to follow the same naming
convention and their call locations were updated accordingly.