Modifies a while_let in a for loop in the hello_can example.
The `hello_can` example function `process` consumes all available transfer, if
any, before returning.

This was done with a `while let Some(...) = iterator.next()`.
As clippy's `while_let_on_iterator` sensibly points out, this is equivalent to a for loop on the iterator,
as it already takes care of stopping when `None` is encountered.

For this reason, the `while let` construct was modified to a simpler `for` construct.