Adds a minimal continuos integration workflow.
As per github's requirements the workflow lives under `.github/workflows/`.

The workflow file was named `ci.yml` and is intended to contain the whole
specification of the continuos integration pipeline.

The minimal file specifies the workflow to run either when a push is made to the
`main` branch or when a pull request that targets the main branch has to be merged.

The first hook is considered required to ensure that main is always in a stable
position.

The repository is currently being developed with a pull-request-first approach,
where every change passes trough a pull-request that is then merged to main.
To ensure the quality of the pull request the second hook was added, running the
workflows for before anything is merged.

While, in a pull-request-first approach, the hook on pushes to `main` may feel
redundant, it is not a given that every push will in effect pass trough a
pull-request or that each pull-request is correctly checked before any merge and
was, thus, considered better to add this second layer of controls to avoid
not noticing problems.

A placeholder `jobs` entry was added that will later be filled with the required
`job` definitions.