Adds a minimal build job.
A new job that builds the project is added, based on the actions-rs family of
github action.

Currently, the job downloads and setups the rust nightly toolchain and builds
all parts of the project the project, using `cargo build` trough the cargo
action, being successful when the build has no errors.