Parametrizes the build job to build for both desktop and embedded.
A strategy matrix was added to condition the job over the target architecture
that we are interested into: x86_64 linux and Cortex-M4F devices.

An additional parameter was added to the matrix to condition what targets to
build.
This is to support the fact that the library's example and tests are expected to
be built only on a std-compatible architecture while the library itself does not
require standard.

Hence, the x86_64 linux architecture builds all available targets, while the
Cortex-M4F architecture only builds the library itself.

The name was conditioned to show the target architecture, such as
Build-x86_64-unknown-linux-gnu for the x86_64 linux build.