Exclude incorrect combination from the Build job matrix.
By default the strategy matrix builds all combination of a job from its
parameter.

In our case, this generates incorrect combinations, such as building all targets
on an embedded device.

To resolve the problem, two exclude directives were added to the build job
matrix to exclude the `--all-targets` - `thumbv7em-none-eabihf` and the `--lib`
- `x86_64-unknown-linux-gnu` combinations.