Corrects the exclusion targets for the build job.
Two embedded targets are currently used for the build job, thumbv7em-none-eabihf
and thumbv6m-none-eabi.

The matrix needs to exclude the `--all-targets` combination for them and,
previously, this was done by having an array in the `exclude` directives
containing both of them.

This seems to not work as expected; in fact it doesn't exclude any of the
required combinations.
The exclude directive probably matches by literal value such that the expressed
exclude would match only when `architecture` is an array of two elements
containing the two architectures.

To resolve the problem, the exclude directive was split into two cases, one for
each architecture triggering for `--all-targets`.