Adds a job that ensure that all automated tests are successful.
The new `Test` job runs cargo's automated tests for the repository, failing when
at least one test is not successful.

For this crate, the tests are written to be run on an std-compatible
environment, and thus are not run on embedded targets.

The job uses `cargo tarpaulin`, trough the `actions-rs/tarpaulin` wrapper, to
further produce code coverage results for the current version of the code.

The coverage is then update to codecov, which has been integrated with the
repository, and uploaded as an artifact for later viewing.

Codecov was chosen as it is simple to integrate into github and provides a free
tier for public repositories. The integration automatically takes care of
reporting coverage changes for PRs and registers metrics about coverage
throughout the repository history.