Adds caching for rust toolchains and cargo target directory.
Makes use of `Swatinem/rust-cache` that automate this process.

The Test job was changed to depend on the Build job so that it makes use of the
cache. Avoiding the need to build the project again and avoiding running when
the project itself cannot be built, which will fail the testing phase without
running tests.