Adds a job to check the formatting of the code.
The new `rustfmt` job runs rustfmt on the whole codebase in check-mode and fails
if there is at least one file that is not completely formatted.

This ensures that consistent formatting is provided throughout the project
lifetime, simplifying diffs and increasing readability to people accustomed to
the project.

The job depends on `build` to avoid running it on an commit which is not compilable.