Adds a job that runs clippy on the codebase.
The new `clippy` job will run `cargo clippy` on the codebase and fail if any
error is encountered.

The job makes use of the `rust-rs/clippy-check` action, which publishes
annotations directly in the code of the commit for each encountered violation.

The job depends on the `build` job, to avoid being run on a codebase that cannot
compile and to make use of the toolchain and build cache.