Adds a job to check that the documentation can be built.
The `documentation` job runs `cargo rustdoc` and check that the documentation
can be built correctly, failing if some error is encountered, such as having
broken references in the documentation.

The job depends on the `build` job to avoid being run on a crate that doesn't
compile and to make use of the toolchain cache.