Avoids panicking when a zero-length frame is received.
`Buildup` assumes that at least 1 byte of payload is present in each frame.

Indeed, a zero-bytes payload is not a uavcan frame as at least the tail-byte
must be present, even when the payload data is empty.

Before this commit, no check was done on the payload of a frame before it was
sent to `Buildup`, this produced, a panic, due to an underflow, when
`Buildup` tried to split the tail-byte from the payload.

To avoid this panic, receive is now modified to check the length of the frame's
payload and rise an error if it is zero.

A test was added to ensure that this is the case.